Created Next Level Wall and Updated Boss Wall script location
