+++ Assets/Scripts/BossWall.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5f2a08021701fd345960562bd6ff3c5b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
