+++ Assets/Scripts/NextLevelWall.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 70ba438c2f1029b4c87cbf83497ad516
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
